Explicitly trigger DiagramSelectionChanged events
So we can trigger the event when a user (re)selects an
already selected element on a diagram.

This is important, since reselecting does not work as expected
after an element has been selected in the model browser.
