+++ gaphor/diagram/tools/__init__.py
@@ -22,7 +22,9 @@ def apply_default_tool_set(view, modeling_language, event_manager, rubberband_st
     view.remove_all_controllers()
     view.add_controller(hover_tool())
     view.add_controller(*text_edit_tools(event_manager))
-    view.add_controller(*transactional_tool(item_tool(), event_manager=event_manager))
+    view.add_controller(
+        *transactional_tool(item_tool(event_manager), event_manager=event_manager)
+    )
     view.add_controller(rubberband_tool(rubberband_state))
     add_basic_tools(view, modeling_language, event_manager)
