+++ gaphor/ui/event.py
@@ -22,13 +22,6 @@ class ModelSelectionChanged:
         self.focused_element = focused_element
 
 
-class DiagramSelectionChanged:
-    def __init__(self, diagram_view, focused_item, selected_items):
-        self.diagram_view = diagram_view
-        self.focused_item = focused_item
-        self.selected_items = selected_items
-
-
 class ToolSelected:
     def __init__(self, tool_name):
         self.tool_name = tool_name
